+++ invaders v2/IShader.h
@@ -15,7 +15,7 @@ public:
 	IShader &operator=(IShader&) = delete;
 	virtual ~IShader(){}
 	virtual void Init(ComPtr<ID3D11Device> device);
-	virtual void RenderShader(ComPtr<ID3D11DeviceContext> context, int indexCount){context->DrawIndexed(indexCount, 0, 0);}
+	virtual void RenderShader(ComPtr<ID3D11DeviceContext> context, size_t indexCount){context->DrawIndexed((UINT)indexCount, 0, 0);}
 protected:
 	void InitializeShader(ComPtr<ID3D11Device> device, wstring vs, wstring ps, const vector<D3D11_INPUT_ELEMENT_DESC> &inputLayout);
 	virtual void InitializeShaderBuffers(ComPtr<ID3D11Device> device);
